Was shown the huge memory cost of the Fullpath on every DirEntry.
Change name of FullPath the test bridge to AlphaFS to GetFullPath to avoid a name clash.
Moved FullPath from DirEntry to CommonEntry class.
Remove obsolete from FlatDirEntryDTO for now.
Moving to only setting Fullpath on IsDirectory DirEntries for now drops 95% of the cost.
Went crazy - created two new enumerators, one returning FlatDirEntryDTO, the other Returning new thing PairDirEntry.  The enumerator for FlatDirEntryDTO has tests but not the PairDirEntry as it was so close to the FDE Enumerator.