Introduced new Flags and BitFields to reduce cost of storing booleans in file and probably memory.
Introduced Hash16 use struct to hold hash so no extra pointer used. (in 64bit the pointer to array of hash cost 8 bytes alone). This should reduce lots and lots of small arrays for hash being needed for holding hashs.
Now using EqualityComparer for DirEntry and Hash16, to avoid the extra arrays of hash that were in use.