Split ICDEWinForm to a file.
Changing Size field in CommonEntry to a long it is just as thats the type from OS anyway.
Added human readable size of a Folder to list views.
Created a helper for drop downs with Display and Value fields.
Created a helper for checkbox with dependent controls being enabled.
Modify DirEntry.SizeCompareWithDirTo() so that the size of the directory which is held in data structure is kept as well.
Implemented the advanced search controls.
Created MyMessageBox that centers on parent to avoid MessageBox.Show().