Fixed loading of cache files crash because there was a dir with name "moo." in current dir.
Bug occurred because load .cde files in current and one folder down.
Using GetFilesWithExtension on a folder with trailing "." fails with an exception.
In this case we don't care, because the intention of loading .cde files one folder down is to allow a ruidmentary level of access control to only people with permission to access them.