Fixed cdeWeb build problem from msbuild with VS2013.
Main problem of build cdeWeb not finding for WebApplications.target is because specwatcher is picking up the .Net version MSBUILD no the one shipped with VS2013. Had to modify water_dot_net.rb and dotnet.watcr.rb - as both have a path hard coded in them.
Reference:
http://stackoverflow.com/questions/17433904/v11-0-webapplications-microsoft-webapplication-targets-was-not-found-when-file-a
http://stackoverflow.com/a/19826448
Switched back the platform with MSBUILD to "Any CPU"